requirements.txt
streamlit
albumentations==1.3.0
config==0.5.1
matplotlib==3.7.1
numpy==1.23.4
pandas==1.3.5
Pillow<10.0.0
plotly==5.15.0
# pytorch_lightning==2.0.6
simple_parsing==0.1.3
streamlit==1.25.0
torch==2.0.1
# torch_pruning==1.2.1
torchmetrics==1.0.1
torchvision==0.15.2
tqdm==4.64.1
wget==3.2
split-folders